Adjust logic for updating the operator
The new logic checks if operator has a value, and if it does, it will
perform the calculation. Then, it assigns the new operator value.

This removes redundancy. The old code would have had to have two
separate operator assignment for when operator is undefined and when it
has a value.